GATT Bearer 'Data OUT' Notifications Re-assembly logic updates
A device that is exposing Mesh Provisioning Service[as a unprovisioned device] or the Mesh Proxy Service [as a provisioned device] could
under-utilize the negotiated ATT MTU and still transmit a valid mesh GATT Proxy PDUs to the peer Proxy Client. The nRF mesh app on Android seems to rely on the negotiated ATT MTU to identify the number of chucks a segmented packet contains. But, whereas, as stated above there can be scenarios where the device exposing the Mesh services under-utilize the ATT MTU. In that case, it will be not possible for the current nRF Android app to Re-assemble the received packet correctly. Yes, this could be very inefficient on the server implementation but the server is not in violation of the protocol by doing so.