Allows for having device MAC Address in the exported JSON.
In a project that we are currently implementing, our nodes are not only LPNs,
But also open the scan window for a bit, to detect nearby LPNs
This allows for social distancing detection, which is a feature much requested during the pandemic.

However, when the scan window is open, only the device mac address is detected, as opposed to the unicast address.
This JSON file provides a viable way to link the device address with the unicast address.